agentic-ai-journey/requirements.txt
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.6
python-dotenv==1.0.0
requests==2.31.0